Day10 part b
Cleverness is required here.
Split the resulting set of diffs into a sequence of 1s
punctuated by 3s. We can't do anything about the 3s; there's
only one way to deal with those.

But the ones we can merge - so we tot up the ways that can
be done, and multiply out the number of approaches for all
the subsequences.