Day10 part b, matrix multiplication style
I'm really rusty on this, but basically the idea here is
that every recurrence relation that's a linear combination
of a fixed set of previous values has a matrix
representation.

By diagonalising the matrix we can rapidly raise it to a
high power and use it to compute arbitrary values of the
sequence.